README.md
For coursework on web technology, 00014714 created this webpage. For my coursework, I have selected the G'uncha Education Center, which is close to my home and it does not have an official website. I mainly utilized HTML, CSS, and JS for this assignment. Throughout the process, I ran into further issues with the page layout, course section, and footer.  I was able to solve these problems after the lecturer’s feedback and online research. All logos and other photos were selected from websites and other information written by myself because it does not have an official website. After completing the full version of this webpage I will approach the authority of the education center and give it to use. I have deployed my website on netlify in URL (00014714.netlify.app). There was one problem like search box toggles existed only on the home page and users can not search data on other pages. I can not find a solution and I will find it as soon as possible. Furthermore, there were some missed pages like FAQ, Testimonials, and Team sections, because I just did not have enough time. I used js to make websites interesting like the navbar toggle, search box, back-to-top button, and so on. Back to top button is a really interesting function, because instead of scrolling up to reach the top of the webpage, the user clicks only one button. I made the buttons work almost all the time. For example, if a user wants to sign up for our mock tests, he just clicks a button and the interface shows the contact section.

Link to my hosting web site: https://00014714.netlify.app